+++ CHANGELOG.md
@@ -39,6 +39,7 @@ on a downstream task.
 - Added a `min_steps` parameter to `BeamSearch` to set a minimum length for the predicted sequences.
 - Added the `FinalSequenceScorer` abstraction to calculate the final scores of the generated sequences in `BeamSearch`. 
 - Added `shuffle` argument to `BucketBatchSampler` which allows for disabling shuffling.
+- Added `allennlp.modules.transformer.attention_module` which contains a generalized `AttentionModule`. `SelfAttention` and `T5Attention` both inherit from this.
 - Added a `Constraint` abstract class to `BeamSearch`, which allows for incorporating constraints on the predictions found by `BeamSearch`,
   along with a `RepeatedNGramBlockingConstraint` constraint implementation, which allows for preventing repeated n-grams in the output from `BeamSearch`.
 - Added `DataCollator` for dynamic operations for each batch.
@@ -53,6 +54,7 @@ on a downstream task.
 - Fixed the potential for a race condition with `cached_path()` when extracting archives. Although the race condition
   is still possible if used with `force_extract=True`.
 - Fixed `wandb` callback to work in distributed training.
+- Fixed `tqdm` logging into multiple files with `allennlp-optuna`.
 
 ## [v2.4.0](https://github.com/allenai/allennlp/releases/tag/v2.4.0) - 2021-04-22